Se crea instancia en la clase Builde
