+++ src/main/java/org/example/creacion/ComidaBuilder.java
@@ -1,0 +1,54 @@
+package org.example.creacion;
+
+public class ComidaBuilder {
+    private Comida comida;
+
+    public ComidaBuilder() {
+
+        comida = new Comida();
+    }
+
+    public ComidaBuilder pan(String pan) {
+        comida.setPan(pan);
+        return this;
+    }
+
+    public ComidaBuilder carne(String carne) {
+        comida.setCarne(carne);
+        return this;
+    }
+
+    public ComidaBuilder lechuga(String lechuga) {
+        comida.setLechuga(lechuga);
+        return this;
+    }
+
+    public ComidaBuilder tomate(String tomate) {
+        comida.setTomate(tomate);
+        return this;
+    }
+
+    public ComidaBuilder queso(String queso) {
+        comida.setQueso(queso);
+        return this;
+    }
+
+    public ComidaBuilder mayonesa(String mayonesa) {
+        comida.setMayonesa(mayonesa);
+        return this;
+    }
+
+    public ComidaBuilder mostaza(String mostaza) {
+        comida.setMostaza(mostaza);
+        return this;
+    }
+
+    public ComidaBuilder ketchup(String ketchup) {
+        comida.setKetchup(ketchup);
+        return this;
+    }
+
+    public Comida build() {
+        return this.comida;
+    }
+}
